aiflow/flink-ai-flow-custom/requirements.txt
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#
# Required dependencies
streamz==0.5.2
matplotlib==3.1.2
grpcio>=1.37.0, <2.0
requests
six
kafka-python==2.0.2
kubernetes==9.0.0
numpy
networkx==2.6.3
oss2==2.15.0
pika==1.1.0
protobuf==3.15.6
pytest==5.3.5
pyyaml>=5.1, <5.4
apscheduler==3.6.3
google-api-core==1.16.0
psutil==5.8.0
pandas
pika==1.1.0
pyecharts==1.8.1
cloudpickle==1.2.2
flask>=1.1.0, <2.0
flask-admin==1.5.4
flask-appbuilder~=3.1.1
flask-caching~=1.9.0
flask_cors>=3.0.10
flask-login>=0.3, <0.5
flask-swagger==0.2.13
flask-wtf>=0.14.2, <0.15
Django~=3.2.7
minio~=7.1.0
# Optional dependencies
apache-flink==1.14.0
tensorflow==2.6.0
pymysql==0.9.3
scikit-learn==0.24.2
alembic>=1.0,<2.0
argcomplete~=1.10
attrs~=20.3.0
cached_property~=1.5
cattrs~=1.1.2
colorlog==4.0.2
configparser>=3.5.0,<3.6.0
croniter>=0.3.17, <0.4
dill>=0.2.2, <0.4
funcsigs>=1.0.0, <2.0.0
future>=0.16.0, <0.19
graphviz>=0.12
gunicorn>=19.5.0, <20.0
iso8601>=0.1.12
jinja2>=2.10.1, <2.11.0
json-merge-patch==0.2
jsonschema~=3.0
lazy_object_proxy~=1.3
markdown>=2.5.2, <3.0
pendulum~=2.0
pygments>=2.0.1, <3.0
python-daemon>=2.1.1, <2.2
python-dateutil>=2.3, <3
setproctitle>=1.1.8, <2
sqlalchemy~=1.3.24
marshmallow-sqlalchemy~=0.23.0
sqlalchemy_jsonfield~=1.0
tabulate>=0.7.5, <0.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift>=0.9.2
typing-extensions>=3.7.4;python_version<"3.8"
tzlocal>=1.4,<2.0.0
unicodecsv>=0.14.1
werkzeug~=1.0
# zope.deprecation>=4.0, <5.0
mysqlclient~=2.0.2
mysql-connector-python==8.0.26
mongoengine~=0.22.1
notification-service==0.2.0
apache-airflow==2.0.0
apache-airflow-providers-ftp==1.1.0
apache-airflow-providers-http==1.1.1
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.2
apache-airflow-providers-celery==1.0.1
redis~=3.2